backend/requirements.txt
# --- Web server & API ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9   # needed for file uploads (Form/File in FastAPI)
httpx>=0.27.0             # Firecrawl HTTP client

# --- Data & DB ---
SQLAlchemy>=2.0.29

# --- Pydantic (FastAPI uses v2) ---
pydantic>=2.6.0

# --- File parsing (PDF/DOCX) ---
pypdf>=4.2.0
python-docx>=1.1.0

# --- LangGraph / LLM glue ---
langgraph>=0.2.27
langchain-core>=0.2.38
langchain-openai>=0.1.7
langchain-groq>=0.1.5    # Added for Groq support
openai>=1.30.0            # backend client used under langchain-openai

# (optional but handy)
python-dotenv>=1.0.1      # if you want to load env vars from .env in backend
langsmith>=0.1.35
langgraph-cli[inmem]>=0.1.46
langserve>=0.2.2
sse-starlette>=1.6.5
langchain_anthropic
e2b